rename file to match pattern
