+++ packages/client/hmi-client/src/components/workflow/ops/model-config/mod.ts
@@ -1,6 +1,6 @@
 import { ModelConfigOperation as operation } from './model-config-operation';
 import node from './tera-model-config-node.vue';
-import drilldown from './tera-model-config.vue';
+import drilldown from './tera-model-config-drilldown.vue';
 
 const name = operation.name;
